+++ src/components/index.ts
@@ -2,5 +2,5 @@ export * from './Elements';
 export * from './Header';
 export * from './Layout';
 export * from './Modals';
-export * from './Selectors';
+export * from './Elements/Selectors';
 export * from './Sidebar';
